# Simplified geoAppendIfWithinShape() and removed spurious calls do sdsdup and sdsfree