In scenarios in which we have large datasets and the elements are not contained within the range we do spurious calls do sdsdup and sdsfree.
I.e. instead of pre-creating an sds before we know if we're gonna use it or not, change the role of geoAppendIfWithinShape to just do geoWithinShape, and let the caller create the string only when needed.

<img width="589" alt="Screenshot 2022-11-17 at 23 43 21" src="https://user-images.githubusercontent.com/5832149/202583542-56e55b3d-4ec6-479f-886d-bb6ba1d01b97.png">

Benchmarking within a geo key with 23244458 elements and the query returns 90K elements, we noticed an improvement from 20.44 ops/sec to 25.01 ops/sec and a drop in the p50 latency from 48.63900 ms to 41.72700 ms. 

To benchmark:

```
memtier_benchmark -c 1 -t 1 --test-time 60 --command "GEOSEARCH key FROMLONLAT 7.0 55.0 BYRADIUS 200 KM" --hide-histogram
``` 

## unstable ( 203b12e41ff7981f0fae5b23819f072d61594813 )

```
ALL STATS
=====================================================================================================
Type            Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
-----------------------------------------------------------------------------------------------------
Geosearchs        20.44        48.93554        48.63900        50.17500        50.43100     24634.70 
Totals            20.44        48.93554        48.63900        50.17500        50.43100     24634.70 
```


## this PR 

```
ALL STATS
=====================================================================================================
Type            Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
-----------------------------------------------------------------------------------------------------
Geosearchs        25.01        39.96836        41.72700        42.23900        43.26300     30147.47 
Totals            25.01        39.96836        41.72700        42.23900        43.26300     30147.47 

```